# oas/openAPI_remediated.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "42C Bank API",
        "description": "An intentionally unsecure banking API for education and training purposes.",
        "contact": {
            "name": "42Crunch Support",
            "url": "https://support.42crunch.com",
            "email": "support@42crunch.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{domain}/apis/banking/{version}",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                },
                "domain": {
                    "default": "localhost:3000",
                    "description": "local development"
                },
                "version": {
                    "enum": [
                        "v1",
                        "v2"
                    ],
                    "default": "v1"
                }
            }
        }
    ],
    "security": [
        {
            "AccessToken": []
        }
    ],
...
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Log in to the account",
                "operationId": "AccountLogin",
                "security": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/Email"
                                    },
                                    "pan": {
                                        "$ref": "#/components/schemas/PersonalAccountNumber"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "access": {
                                            "type": "string",
                                            "pattern": "^eyJ[A-Za-z0-9\\-_]+.eyJ[A-Za-z0-9\\-_]+.[A-Za-z0-9\\-_+]*$",
                                            "maxLength": 500
                                        },
                                        "access-expires": {
                                            "type": "number",
                                            "maximum": 3600
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Email": {
                "type": "string",
                "format": "email-secured",
                "example": "jbernard@email.com",
                "pattern": "^\\w+(?:[\\-.]\\w+)*@[0-9A-Za-z]+(?:[\\-.][0-9A-Za-z]+)*\\.(?:[A-Za-z][0-9A-Za-z]+|[0-9]+[A-Za-z][0-9A-Za-z]*)$",
                "minLength": 6,
                "maxLength": 320,
                "x-42c-format": "o:email-secured"
            },
...
            "PersonalAccountNumber": {
                "type": "string",
                "description": "personal account number",
                "pattern": "^[0-9]{5,5}$",
                "example": "33333"
            },
            "ResponseMessage": {
                "type": "string",
                "pattern": "^[A-Za-z0-9\/\\.\"', ]{1,100}$"
            },
            "MessageObject": {
                "type": "object",
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/ResponseMessage"
                    }
                },
                "additionalProperties": false
            },
...
        },
        "responses": {
...
            "406ErrorResponse": {
                "description": "Not Acceptable",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
            "415ErrorResponse": {
                "description": "Unsupported Media Type",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
            "429ErrorResponse": {
                "description": "Too Many Requests",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
            "DefaultErrorResponse": {
                "description": "Unexpected Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "AccessToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}